Extract name search form to component
