+++ packages/name-service/src/App/routes/Contract/components/SearchResult/SearchResult.less
